Additional protections against mutating customer data
Separated out a “fetch” command that interacts with customer data. It is the only command that is allowed to connect to the PROD BridalLive environment. All other commands are only allowed to run against the BridalLive QA environment.

This was done by routing all node-fetch calls in `src/integrations/BridalLive/api.ts` through a safeFetch function that prevents requests to BL PROD (https://app.bridallive.com) unless a global ‘isCustomerAction’ boolean is set to true. The global ‘isCustomerAction’ boolean is only set to  true by the ‘fetch’ command.

This security measure is in addition to the previously existing code that limits data mutations to an authenticated token that belongs to the expected “demo” store defined in `src/settings.ts`.