Fetch/import all items, not just gowns
Since we might use this demo environment for testing, the data should reflect a real customer environment as much as possible. For each item imported, the dept gets set to the demo Bridal Gown dept or the demo Other dept (so we don’t have to maintain multiple departments in the demo env).

Also, includes population code for Receiving Vouchers and associated RV Line Items.